Move readonly prop group into its own component
